[style/#21] - chat.ts -> chat.type.ts 변경
